1. Refactored all test cases and removed some code duplications 2. Added logback library 3. Added allure reporting tool

+++ SeleniumProject/src/test/java/testngtests/AdminPanelTest.java
@@ -2,18 +2,19 @@ package testngtests;
 
 import com.mailosaur.MailosaurException;
 import models.User;
-import models.testngpages.AdminPanelPage;
-import org.testng.Assert;
+import models.testngpages.adminpanel.AdminPanelPage;
+import org.assertj.core.api.Assertions;
 import org.testng.annotations.Test;
 import testngtests.abstractclasses.AbstractTest;
 
 import java.io.IOException;
 
+import static models.enums.SystemAlerts.*;
+
 public class AdminPanelTest extends AbstractTest {
 
     private User newUser = new User.UserBuilder().build();
-    private static final String SUCCESS_ALERT_PROMOTING_USER_TEXT = "Success! User with email '%s' has been promoted to ADMIN role.";
-    private static final String SUCCESS_ALERT_DEMOTING_USER_TEXT = "Success! User with email '%s' has been demoted to USER role.";
+
     private AdminPanelPage adminPanelPage;
 
     @Test
@@ -22,41 +23,71 @@ public class AdminPanelTest extends AbstractTest {
 
         signIn(existedAdminUser);
         adminPanelPage = mainPage.goToAdminPanel();
-        Assert.assertTrue(adminPanelPage.isPromoteButtonAvailableForUserWithEmail(newUser.getEmail()));
-        Assert.assertFalse(adminPanelPage.isDemoteButtonAvailableForUserWithEmail(newUser.getEmail()));
-        Assert.assertTrue(adminPanelPage.isDeleteButtonAvailableForUserWithEmail(newUser.getEmail()));
+        assertThatPromoteButtonIsAvailableOrNot(true);
+        assertThatDemoteButtonIsAvailableOrNot(false);
+        assertThatDeleteButtonIsAvailableOrNot(true);
         adminPanelPage.promoteUserByEmail(newUser.getEmail());
-        Assert.assertEquals(adminPanelPage.getTextFromSuccessAlert(), String.format(SUCCESS_ALERT_PROMOTING_USER_TEXT, newUser.getEmail()));
-        Assert.assertFalse(adminPanelPage.isPromoteButtonAvailableForUserWithEmail(newUser.getEmail()));
-        Assert.assertTrue(adminPanelPage.isDemoteButtonAvailableForUserWithEmail(newUser.getEmail()));
-        Assert.assertTrue(adminPanelPage.isDeleteButtonAvailableForUserWithEmail(newUser.getEmail()));
+        assertThatSuccessAlertHasExpectedText(String.format(SUCCESS_ALERT_PROMOTING_USER_TEXT.getAlertText(), newUser.getEmail()));
+        assertThatPromoteButtonIsAvailableOrNot(false);
+        assertThatDemoteButtonIsAvailableOrNot(true);
+        assertThatDeleteButtonIsAvailableOrNot(true);
         signOut();
 
         signIn(newUser);
-        Assert.assertTrue(mainPage.isAdminPanelButtonVisible());
+        assertThatAdminPanelButtonIsAvailableOrNot(true);
         signOut();
 
         signIn(existedAdminUser);
         adminPanelPage = mainPage.goToAdminPanel();
         adminPanelPage.demoteUserByEmail(newUser.getEmail());
-        Assert.assertEquals(adminPanelPage.getTextFromSuccessAlert(), String.format(SUCCESS_ALERT_DEMOTING_USER_TEXT, newUser.getEmail()));
-        Assert.assertTrue(adminPanelPage.isPromoteButtonAvailableForUserWithEmail(newUser.getEmail()));
-        Assert.assertFalse(adminPanelPage.isDemoteButtonAvailableForUserWithEmail(newUser.getEmail()));
-        Assert.assertTrue(adminPanelPage.isDeleteButtonAvailableForUserWithEmail(newUser.getEmail()));
+        assertThatSuccessAlertHasExpectedText(String.format(SUCCESS_ALERT_DEMOTING_USER_TEXT.getAlertText(), newUser.getEmail()));
+        assertThatPromoteButtonIsAvailableOrNot(true);
+        assertThatDemoteButtonIsAvailableOrNot(false);
+        assertThatDeleteButtonIsAvailableOrNot(true);
         signOut();
 
         signIn(newUser);
-        Assert.assertFalse(mainPage.isAdminPanelButtonVisible());
+        assertThatAdminPanelButtonIsAvailableOrNot(false);
         signOut();
 
         signIn(existedAdminUser);
         adminPanelPage = mainPage.goToAdminPanel();
         adminPanelPage.deleteUserByEmail(newUser.getEmail());
-        Assert.assertEquals(adminPanelPage.getTextFromSuccessAlert(), SUCCESS_ALERT_USER_REMOVED_TEXT);
-        Assert.assertFalse(adminPanelPage.isRowForUserWithEmailExist(newUser.getEmail()));
+        assertThatSuccessAlertHasExpectedText(SUCCESS_ALERT_USER_REMOVED_TEXT.getAlertText());
+        assertThatRowWithUserIsNotExist(newUser);
         signOut();
 
         signIn(newUser, true);
+    }
+
+    private void assertThatPromoteButtonIsAvailableOrNot(boolean shouldBeAvailable) {
+        Assertions.assertThat(adminPanelPage.isPromoteButtonAvailableForUserWithEmail(newUser.getEmail()))
+                .as("Assert that Promote button is available: " + shouldBeAvailable)
+                .isEqualTo(shouldBeAvailable);
+    }
+
+    private void assertThatDemoteButtonIsAvailableOrNot(boolean shouldBeAvailable) {
+        Assertions.assertThat(adminPanelPage.isDemoteButtonAvailableForUserWithEmail(newUser.getEmail()))
+                .as("Assert that Demote button is available: " + shouldBeAvailable)
+                .isEqualTo(shouldBeAvailable);
+    }
+
+    private void assertThatDeleteButtonIsAvailableOrNot(boolean shouldBeAvailable) {
+        Assertions.assertThat(adminPanelPage.isDeleteButtonAvailableForUserWithEmail(newUser.getEmail()))
+                .as("Assert that Delete button is available: " + shouldBeAvailable)
+                .isEqualTo(shouldBeAvailable);
+    }
+
+    private void assertThatAdminPanelButtonIsAvailableOrNot(boolean shouldBeAvailable) {
+        Assertions.assertThat(adminPanelPage.isAdminPanelButtonVisible())
+                .as("Assert that Admin panel button is available: " + shouldBeAvailable)
+                .isEqualTo(shouldBeAvailable);
+    }
+
+    private void assertThatRowWithUserIsNotExist(User user) {
+        Assertions.assertThat(adminPanelPage.isRowForUserWithEmailExist(user.getEmail()))
+                .as("Assert that row for expected user is not exist")
+                .isFalse();
     }
 
 }
